daemon: move download `jsonrpc_` methods to an individual module
The base methods are moved to new modules `daemon_meta`
and `daemon_get` so that the methods in Daemon aren't placed
all in the same file.

The big class `Daemon` is now derived from the new `Daemon_get`
instead of the original metaclass `JSONRPCServerType`.
```
JSONRPCServerType -> Daemon_get -> Daemon
```
